requirements.txt
beautifulsoup4==4.9.3
matplotlib==3.3.4
python-dotenv==0.15.0
requests==2.25.1
scipy==1.5.4
selenium==3.141.0